eval.rs: remove is_test and embed mock-nixpkgs.nix
This commit does the following things:

1. Purify the environment variables entirely when calling `nix-instantiate`.
   Opt in to the five environment variables needed in order to run
   `nix-instantiate` in the Nix sandbox.
2. Remove the hard-coded `NIX_PATH` environment variables in `default.nix`
   and `package.nix`. These were implicitly sent through to `nix-instantiate`,
   along with `-I` arguments. Now, only `-I` is used.
3. Replace the `is_test` argument with `#[cfg(test)]` applied to
   `mutate_nix_instatiate_arguments_based_on_cfg`.

While this is slightly more lines of code that the combination of Nix and Rust
that it replaces, there is less Nix logic and more Rust logic.

Additionally, the Rust logic that is test-only is contained to `cargo check`
only rather than being present in the production binary.