fix autoPrUpdate scripts
It seems that we now need an explicit dependency on nix to get nix-prefetch-url.

It also seems that openssh is required for git to work
(as is required by npins and cargo), at least locally for me.
This may be due to my local configuration?
Even if it is, I think there's minimal harm in adding the requirement.